requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudevents==1.2.0
colorama==0.4.4
deprecation==2.1.0
et-xmlfile==1.1.0
firebase-admin==5.2.0
Flask==2.0.3
frozenlist==1.3.0
functions-framework==3.0.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
line-bot-sdk==2.1.0
MarkupSafe==2.1.1
msgpack==1.0.3
multidict==6.0.2
numpy==1.21.5
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
watchdog==1.0.2
Werkzeug==2.0.3
yarl==1.7.2
